Move EventEmitter
